parse idx header

+++ mnist.h
@@ -1,6 +1,7 @@
 #ifndef MNIST_H
 #define MNIST_H
 #include "matrix.h"
+#include <stdint.h>
 
 #define MNIST_TRAINING_IMAGE_FILE "mnist/train-images.idx3-ubyte"
 #define MNIST_TRAINING_LABEL_FILE "mnist/train-labels.idx1-ubyte"
@@ -8,10 +9,10 @@
 #define MNIST_TEST_LABEL_FILE "mnist/t10k-labels.idx1-ubyte"
 
 typedef struct {
-    int size_training;
-    int size_test;
-    int * training_labels;
-    int * test_labels;
+    uint32_t size_training;
+    uint32_t size_test;
+    uint8_t * training_labels;
+    uint8_t * test_labels;
     Matrix * training_images;
     Matrix * test_images;
 } MNISTData;
